fix(blocks-aggird): Apply to aggrid block.
